+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+- Feat: Add publish command to cli
+
 ## [0.4.0] - 2022-11-01
 
 - Feat: Add an option to get version from distribution metadata